+++ src/js/route.js
@@ -28,8 +28,8 @@ class Route {
         // If  we're building just a path there's no origin, otherwise: if this route has a
         // domain configured we construct the origin with that, if not we use the app URL
         const origin = !this.config.absolute ? '' : this.definition.domain
-            ? `${this.config.baseUrl.match(/^\w+:\/\//)[0]}${this.definition.domain}${this.config.basePort ? `:${this.config.basePort}` : ''}`
-            : this.config.baseUrl;
+            ? `${this.config.url.match(/^\w+:\/\//)[0]}${this.definition.domain}${this.config.port ? `:${this.config.port}` : ''}`
+            : this.config.url;
 
         return `${origin}/${this.definition.uri}`;
     }
@@ -104,11 +104,11 @@ class Router extends String {
         this._config = config ?? Ziggy ?? globalThis?.Ziggy;
 
         if (name) {
-            if (!this._config.namedRoutes[name]) {
+            if (!this._config.routes[name]) {
                 throw new Error(`Ziggy error: route '${name}' is not in the route list.`);
             }
 
-            this._route = new Route(name, this._config.namedRoutes[name], { ...this._config, absolute });
+            this._route = new Route(name, this._config.routes[name], { ...this._config, absolute });
             this._params = this._parse(params);
         }
     }
@@ -157,7 +157,7 @@ class Router extends String {
         const url = window.location.host + window.location.pathname;
 
         // Find the first route that matches the current URL
-        const [current, route] = Object.entries(this._config.namedRoutes).find(
+        const [current, route] = Object.entries(this._config.routes).find(
             ([_, route]) => new Route(name, route, this._config).matchesUrl(url)
         );
 
@@ -189,7 +189,7 @@ class Router extends String {
      * @return {Object}
      */
     get params() {
-        return this._dehydrate(this._config.namedRoutes[this.current()]);
+        return this._dehydrate(this._config.routes[this.current()]);
     }
 
     /**
@@ -199,7 +199,7 @@ class Router extends String {
      * @return {Boolean}
      */
     has(name) {
-        return Object.keys(this._config.namedRoutes).includes(name);
+        return Object.keys(this._config.routes).includes(name);
     }
 
     /**
@@ -221,7 +221,7 @@ class Router extends String {
         params = ['string', 'number'].includes(typeof params) ? [params] : params;
 
         // Separate segments with and without defaults, and fill in the default values
-        const segments = route.parameterSegments.filter(({ name }) => !this._config.defaultParameters[name]);
+        const segments = route.parameterSegments.filter(({ name }) => !this._config.defaults[name]);
 
         if (Array.isArray(params)) {
             // If the parameters are an array they have to be in order, so we can transform them into
@@ -256,8 +256,8 @@ class Router extends String {
      * @return {Object} Default route parameters.
      */
     _defaults(route) {
-        return route.parameterSegments.filter(({ name }) => this._config.defaultParameters[name])
-            .reduce((result, { name }, i) => ({ ...result, [name]: this._config.defaultParameters[name] }), {});
+        return route.parameterSegments.filter(({ name }) => this._config.defaults[name])
+            .reduce((result, { name }, i) => ({ ...result, [name]: this._config.defaults[name] }), {});
     }
 
     /**
@@ -304,7 +304,7 @@ class Router extends String {
     _dehydrate(route) {
         let pathname = window.location.pathname
             // If this Laravel app is in a subdirectory, trim the subdirectory from the path
-            .replace(this._config.baseUrl.replace(/^\w*:\/\/[^/]+/, ''), '')
+            .replace(this._config.url.replace(/^\w*:\/\/[^/]+/, ''), '')
             .replace(/^\/+/, '');
 
         // Given part of a valid 'hydrated' URL containing all its parameter values,